requirements.txt
fastapi
uvicorn
sse-starlette
ddgs
google-generativeai
openai
python-dotenv
google-api-python-client
